// Desktop/client-mis/types/types.ts
export interface IProductCart {
  id: number;
  product: {
    id: number;
    brand: {
      id: number;
      image: string;
      name: string;
    };
    color: {
      id: number;
      name: string;
      color: string;
    };
    size: {
      id: number;
      name: string;
    };
    manufacturerCountry: {
      id: number;
      name: string;
    };
    subTypes: {
      id: number;
      name: string;
      type: number;
    }[];
    name: string;
    description: string;
    image: string;
    article_number: string;
    amount: number;
    compound: string;
    price: string;
    season: string;
    pattern: string;
    slug: string;
    is_famous: boolean;
    date_and_time: string;
  };
  amount: number;
  date: string;
  order: number;
}

export interface IProductMainPage {
  id: number;
  brand: {
    id: number;
    image: string;
    name: string;
  };
  color: {
    id: number;
    name: string;
    color: string;
  };
  size: {
    id: number;
    name: string;
  };
  manufacturerCountry: {
    id: number;
    name: string;
  };
  subTypes: {
    id: number;
    name: string;
    type: number;
  }[];
  first_image: string;
  second_image: string;
  third_image: string;
  fourth_image: string;
  fifth_image: string;
  name: string;
  description: string;
  article_number: string;
  amount: number;
  compound: string;
  price: string;
  season: string;
  pattern: string;
  slug: string;
  is_famous: boolean;
  date_and_time: string;
}